Make both maxsize and eotmargin 'size_t' so that cygwin doesn't complain about comparing signed vs unsigned
git-svn-id: file:///home/jj/hercules.svn/trunk@1629 956126f8-22a0-4046-8f4a-272fa8102e63

+++ hercules.h
@@ -1181,7 +1181,7 @@ typedef struct _DEVBLK {
                                         /* MVS 3.8j workaround       */
         U16 logical_readonly;       /* Tape is forced READ ONLY  */
             U16 chksize;                /* Chunk size                */
-            off_t maxsize;              /* Maximum allowed TAPE file
+            size_t maxsize;              /* Maximum allowed TAPE file
                                            size                      */
             size_t eotmargin;           /* Amount of space left
                                            before reporting EOT      */
